job-execution-service/requirements.txt
# Job Execution Service Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Celery and Redis
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# SSH and Remote Connections
asyncssh==2.14.0
# pywinrm==0.4.3  # Uncomment when implementing WinRM

# Scheduling
croniter==2.0.1

# File Operations
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1